Work on Restructuring the Data Schema
This contains a number of structural changes.  Most notably, it adds card modes to the card Type property and removes property elements from set.xml files if they are empty to reduce extraneous content.  Larger changes were made on the backend used to create the package.

Other changes include fixing some data errors and updating the card property "Trait" to be "Traits."  Additionally, the code for switching card modes has been completely rewritten and given a new name (switchModes).

Given the extent of the changes, should be mindful of issues/errors creeping in.