Tidying Up Some Code/Minor Changes
Small tidying up by fixing some non-standard indents and removing a superfluous try/except wrapper.
Fixing the way findCharacter searches for characters to make it work with the recent addition of subtype in the card.type proeprty.
Making the clearTarget function remove all targeting, now just targeting on your own cards.  Not ideal, but it is an improvement. May want to revisit this later.
Updating the pickBody function to try to make it work properly with the new subtyping in card.type.