Preliminary Upgrade Alignment
A new alignment system for the package would be desirable, but progress on it has stalled due to the complexity.  For now, adding in a placeholder function for playing Upgrades on bots in a semi-orderly mannered.  Doesn't handle multiple upgrades of one type, and it's preliminary so the aesthetics of the layout can be improved.

Also changing capitalization on a dialog box title to standardize.