Overhaul to Play Upgrade
Created an upgrade management system to let OCTGN handle upgrades more intelligently.  The crucial factor is using the new custom card property, filler, to track which character upgrades are equipped to.

The use of % and // is to place upgrades in an arbitrary number of rows in three columns.  This is more flexible than the previous approach, especially since the ability to rearrange cards within the dialog box will allow users to change how upgrades are displayed.