Fixing a Notification Message
For some reason a the notification message given when trying to flip a card with only one side informs the players' what side the card is currently on.

Also switching it to a whisper instead of notify as there's no need to tell the opponent a player has attempted to flip a card that has only one side.