Minor Optimization to Notification Message
A notification message in the switchModes() function uses the Python split function to extract what mode a card is being flipped to for display purposes.  It was written to assume the text we want to display would come after a hyphen.*  This assumption meant if a card type did not have a hyphen, an error would be thrown.

Changing the list index from [1] to [-1] tells Python to extract the last string which follows a hyphen, defaulting to the full string if no hyphen exists.  This will functionally be the same in cases with 1 hyphen but won't thrown an error in cases with 0 hyphens.  It will also change the functionality in cases with 2+ hyphens, though neither approach will produce intelligible results in those cases.

Practically speaking this should have no effect as the code should never run on cards whose card type does not have a hyphen.  However, there is no downside and it has the upside of handling any potential 0-hyphen cases more cleanly.

*All mentions of "hyphen" should be understood to mean "- " with a whitespace following a hyphen.